Move track page lineup reducer to common
