+++ src/common/store/pages/track/lineup/reducer.js
